Optimize container startup
The operations to load KeyedVectors from word2vec data, calculate the linear transform, and compile the training data vectors are very slow, but the result is the same every time.  So instead of incurring this large cost on every container start, we can pre-compile the various vectors as part of the image build and store just the compiled representation in the final runtime image.  This reduces the cold startup time for containers from several minutes to around 15-20 seconds.
