+++ buildscripts/w2v_to_kv.py
@@ -1,0 +1,5 @@
+from gensim.models import KeyedVectors
+import sys
+
+model = KeyedVectors.load_word2vec_format(sys.argv[1])
+model.save(sys.argv[2])
